@@ -1,12 +1,13 @@
 name: kdl
-version: 1.0.1
-description: Dart implementation of the KDL Document Language
+version: 2.0.0
+description: KDL is a small, pleasant document language with XML-like node semantics that looks like you're invoking a bunch of CLI commands!
 repository: https://github.com/danini-the-panini/kdl-dart
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
 dependencies:
   big_decimal: ^0.5.0
-  string_scanner: ^1.1.1
+  string_scanner: ^1.2.0
 dev_dependencies:
-  test: ^1.21.4
-  path: ^1.8.0
+  test: ^1.25.2
+  path: ^1.9.0
+  lints: ^5.1.1
