@@ -1,15 +1,14 @@
-import 'package:big_decimal/big_decimal.dart';
 import 'package:test/test.dart';
 
 import 'package:kdl/src/document.dart';
-import 'package:kdl/src/parser.dart';
 import 'package:kdl/src/exception.dart';
+import 'package:kdl/src/parser.dart';
 
 void main() {
-  late KdlParser parser;
+  late KdlV1Parser parser;
 
   setUp(() {
-    parser = KdlParser();
+    parser = KdlV1Parser();
   });
 
   test('parse_empty_string', () {
@@ -34,7 +33,7 @@ void main() {
           KdlNode('node', arguments: [KdlInt(1)])
         ])));
     expect(
-        parser.parse('node 1 2 "3" #true #false #null'),
+        parser.parse('node 1 2 "3" true false null'),
         equals(KdlDocument([
           KdlNode('node', arguments: [
             KdlInt(1),
@@ -51,24 +50,9 @@ void main() {
           KdlNode('node', children: [KdlNode('node2')])
         ])));
     expect(
-        parser.parse("node {\n    node2    \n}"),
-        equals(KdlDocument([
-          KdlNode('node', children: [KdlNode('node2')])
-        ])));
-    expect(
         parser.parse("node { node2; }"),
         equals(KdlDocument([
           KdlNode('node', children: [KdlNode('node2')])
-        ])));
-    expect(
-        parser.parse("node { node2 }"),
-        equals(KdlDocument([
-          KdlNode('node', children: [KdlNode('node2')])
-        ])));
-    expect(
-        parser.parse("node { node2; node3 }"),
-        equals(KdlDocument([
-          KdlNode('node', children: [KdlNode('node2'), KdlNode('node3')])
         ])));
   });
 
@@ -77,7 +61,7 @@ void main() {
     expect(parser.parse('/- node'), equals(KdlDocument([])));
     expect(parser.parse("/- node\n"), equals(KdlDocument([])));
     expect(parser.parse('/-node 1 2 3'), equals(KdlDocument([])));
-    expect(parser.parse('/-node key=#false'), equals(KdlDocument([])));
+    expect(parser.parse('/-node key=false'), equals(KdlDocument([])));
     expect(parser.parse("/-node{\nnode\n}"), equals(KdlDocument([])));
     expect(parser.parse("/-node 1 2 3 key=\"value\" \\\n{\nnode\n}"),
         equals(KdlDocument([])));
@@ -137,24 +121,14 @@ void main() {
           KdlNode('node', arguments: [KdlString("hello\nworld")])
         ])));
     expect(
-        parser.parse(r'node -flag'),
-        equals(KdlDocument([
-          KdlNode('node', arguments: [KdlString("-flag")])
-        ])));
-    expect(
-        parser.parse(r'node --flagg'),
-        equals(KdlDocument([
-          KdlNode('node', arguments: [KdlString("--flagg")])
-        ])));
-    expect(
         parser.parse(r'node "\u{10FFF}"'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString("\u{10FFF}")])
         ])));
     expect(
-        parser.parse(r'node "\"\\\b\f\n\r\t"'),
+        parser.parse(r'node "\"\\\/\b\f\n\r\t"'),
         equals(KdlDocument([
-          KdlNode('node', arguments: [KdlString("\"\\\u{08}\u{0C}\n\r\t")])
+          KdlNode('node', arguments: [KdlString("\"\\/\u{08}\u{0C}\n\r\t")])
         ])));
     expect(
         parser.parse(r'node "\u{10}"'),
@@ -163,32 +137,10 @@ void main() {
         ])));
     expect(() {
       parser.parse(r'node "\i"');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse(r'node "\u{c0ffee}"');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse(r'node "oops');
-    }, throwsA(isA<KdlParseException>()));
-  });
-
-  test('unindented multiline strings', () {
-    expect(
-        parser.parse('node """\n  foo\n  bar\n    baz\n  qux\n  """'),
-        equals(KdlDocument([
-          KdlNode('node', arguments: [KdlString("foo\nbar\n  baz\nqux")])
-        ])));
-    expect(
-        parser.parse('node #"""\n  foo\n  bar\n    baz\n  qux\n  """#'),
-        equals(KdlDocument([
-          KdlNode('node', arguments: [KdlString("foo\nbar\n  baz\nqux")])
-        ])));
-    expect(() {
-      parser.parse('node """\n    foo\n  bar\n    baz\n    """');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node #"""\n    foo\n  bar\n    baz\n    """#');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
   });
 
   test('float', () {
@@ -233,20 +185,20 @@ void main() {
           KdlNode('node', arguments: [KdlBigDecimal.from(123456789.0)])
         ])));
     expect(() {
+      parser.parse('node ?1.0');
+    }, throwsA(anything));
+    expect(() {
+      parser.parse('node _1.0');
+    }, throwsA(anything));
+    expect(() {
       parser.parse('node 1._0');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 1.');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node 1.0v2');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node -1em');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node .0');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
   });
 
   test('integer', () {
@@ -280,6 +232,18 @@ void main() {
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlInt(-123456789)])
         ])));
+    expect(() {
+      parser.parse('node ?0123456789');
+    }, throwsA(anything));
+    expect(() {
+      parser.parse('node _0123456789');
+    }, throwsA(anything));
+    expect(() {
+      parser.parse('node a');
+    }, throwsA(anything));
+    expect(() {
+      parser.parse('node --');
+    }, throwsA(anything));
   });
 
   test('hexadecimal', () {
@@ -300,13 +264,13 @@ void main() {
         ])));
     expect(() {
       parser.parse('node 0x_123');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0xg');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0xx');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
   });
 
   test('octal', () {
@@ -327,13 +291,13 @@ void main() {
         ])));
     expect(() {
       parser.parse('node 0o_123');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0o8');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0oo');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
   });
 
   test('binary', () {
@@ -359,64 +323,56 @@ void main() {
         ])));
     expect(() {
       parser.parse('node 0b_0110');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0b20');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
     expect(() {
       parser.parse('node 0bb');
-    }, throwsA(isA<KdlParseException>()));
+    }, throwsA(anything));
   });
 
   test('raw_string', () {
     expect(
-        parser.parse(r'node #"foo"#'),
+        parser.parse(r'node r"foo"'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString('foo')])
         ])));
     expect(
-        parser.parse(r'node #"foo\nbar"#'),
+        parser.parse(r'node r"foo\nbar"'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString(r'foo\nbar')])
         ])));
     expect(
-        parser.parse(r'node #"foo"#'),
+        parser.parse(r'node r#"foo"#'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString('foo')])
         ])));
     expect(
-        parser.parse(r'node ##"foo"##'),
+        parser.parse(r'node r##"foo"##'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString('foo')])
         ])));
     expect(
-        parser.parse(r'node #"\nfoo\r"#'),
+        parser.parse(r'node r"\nfoo\r"'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlString(r'\nfoo\r')])
         ])));
     expect(() {
-      parser.parse('node ##"foo"#');
-    }, throwsA(isA<KdlParseException>()));
+      parser.parse('node r##"foo"#');
+    }, throwsA(anything));
   });
 
   test('boolean', () {
     expect(
-        parser.parse('node #true'),
+        parser.parse('node true'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlBool(true)])
         ])));
     expect(
-        parser.parse('node #false'),
+        parser.parse('node false'),
         equals(KdlDocument([
           KdlNode('node', arguments: [KdlBool(false)])
-        ])));
-  });
-
-  test('null', () {
-    expect(
-        parser.parse('node #null'),
-        equals(KdlDocument([
-          KdlNode('node', arguments: [KdlNull()])
         ])));
   });
 
@@ -462,22 +418,13 @@ void main() {
 
   test('escline', () {
     expect(
-        parser.parse("node\\\n  1"),
+        parser.parse("foo\\\n  1"),
         equals(KdlDocument([
-          KdlNode('node', arguments: [KdlInt(1)])
+          KdlNode('foo', arguments: [KdlInt(1)])
         ])));
-    expect(parser.parse("node\\\n"), equals(KdlDocument([KdlNode('node')])));
-    expect(parser.parse("node\\ \n"), equals(KdlDocument([KdlNode('node')])));
-    expect(parser.parse("node\\\n "), equals(KdlDocument([KdlNode('node')])));
     expect(() {
-      parser.parse('node \\foo');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node\\\\\nnode2');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node \\\\\nnode2');
-    }, throwsA(isA<KdlParseException>()));
+      parser.parse("node\\\nnode2");
+    }, throwsA(anything));
   });
 
   test('whitespace', () {
@@ -498,160 +445,6 @@ void main() {
         equals(KdlDocument([KdlNode('node1'), KdlNode('node2')])));
   });
 
-  test('basic', () {
-    var doc = parser.parse('title "Hello, World"');
-    var nodes = KdlDocument([
-      KdlNode('title', arguments: [KdlString("Hello, World")]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('multiple values', () {
-    var doc = parser.parse('bookmarks 12 15 188 1234');
-    var nodes = KdlDocument([
-      KdlNode('bookmarks',
-          arguments: [KdlInt(12), KdlInt(15), KdlInt(188), KdlInt(1234)]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('properties', () {
-    var doc = parser.parse("""
-author "Alex Monad" email="alex@example.com" active= #true
-foo bar =#true "baz" quux =\\
-  #false 1 2 3
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode(
-        'author',
-        arguments: [KdlString("Alex Monad")],
-        properties: {
-          'email': KdlString("alex@example.com"),
-          'active': KdlBool(true),
-        },
-      ),
-      KdlNode(
-        'foo',
-        arguments: [KdlString("baz"), KdlInt(1), KdlInt(2), KdlInt(3)],
-        properties: {
-          'bar': KdlBool(true),
-          'quux': KdlBool(false),
-        },
-      ),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('nested child nodes', () {
-    var doc = parser.parse("""
-contents {
-  section "First section" {
-    paragraph "This is the first paragraph"
-    paragraph "This is the second paragraph"
-  }
-}
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('contents', children: [
-        KdlNode('section', arguments: [
-          KdlString("First section")
-        ], children: [
-          KdlNode('paragraph',
-              arguments: [KdlString("This is the first paragraph")]),
-          KdlNode('paragraph',
-              arguments: [KdlString("This is the second paragraph")]),
-        ]),
-      ]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('semicolon', () {
-    var doc = parser.parse("node1; node2; node3;");
-    var nodes = KdlDocument([
-      KdlNode('node1'),
-      KdlNode('node2'),
-      KdlNode('node3'),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('optional child semicolon', () {
-    var doc = parser.parse('node {foo;bar;baz}');
-    var nodes = KdlDocument([
-      KdlNode('node', children: [
-        KdlNode('foo'),
-        KdlNode('bar'),
-        KdlNode('baz'),
-      ]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('raw strings', () {
-    var doc = parser.parse("""
-node "this\\nhas\\tescapes"
-other #"C:\\Users\\zkat\\"#
-other-raw #"hello"world"#
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('node', arguments: [KdlString("this\nhas\tescapes")]),
-      KdlNode('other', arguments: [KdlString("C:\\Users\\zkat\\")]),
-      KdlNode('other-raw', arguments: [KdlString("hello\"world")]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('multiline strings', () {
-    var doc = parser.parse("""
-string \"""
-my
-multiline
-value
-\"""
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('string', arguments: [KdlString("my\nmultiline\nvalue")]),
-    ]);
-    expect(doc, equals(nodes));
-
-    expect(() {
-      parser.parse('node """foo"""');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node #"""foo"""#');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node """\n  oops');
-    }, throwsA(isA<KdlParseException>()));
-    expect(() {
-      parser.parse('node #"""\n  oops');
-    }, throwsA(isA<KdlParseException>()));
-  });
-
-  test('numbers', () {
-    var doc = parser.parse("""
-      num 1.234e-42
-      my-hex 0xdeadbeef
-      my-octal 0o755
-      my-binary 0b10101101
-      bignum 1_000_000
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('num', arguments: [KdlBigDecimal(BigDecimal.parse('1.234e-42'))]),
-      KdlNode('my-hex', arguments: [KdlInt(0xdeadbeef)]),
-      KdlNode('my-octal', arguments: [KdlInt(493)]),
-      KdlNode('my-binary', arguments: [KdlInt(173)]),
-      KdlNode('bignum', arguments: [KdlInt(1000000)]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
   test('comments', () {
     var doc = parser.parse("""
       // C style
@@ -660,7 +453,7 @@ value
       C style multiline
       */
 
-      tag /*foo=#true*/ bar=#false
+      tag /*foo=true*/ bar=false
 
       /*/*
       hello
@@ -673,47 +466,10 @@ value
     expect(doc, equals(nodes));
   });
 
-  test('slash dash', () {
-    var doc = parser.parse("""
-/-mynode "foo" key=1 {
-  a
-  b
-  c
-}
-
-mynode /- "commented" "not commented" /-key="value" /-{
-  a
-  b
-}
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('mynode', arguments: [KdlString("not commented")]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('multiline nodes', () {
-    var doc = parser.parse("""
-title \\
-  "Some title"
-
-my-node 1 2 \\  // comments are ok after \\
-        3 4
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('title', arguments: [KdlString("Some title")]),
-      KdlNode('my-node',
-          arguments: [KdlInt(1), KdlInt(2), KdlInt(3), KdlInt(4)]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
   test('utf8', () {
     var doc = parser.parse("""
-smile "😁"
-ノード お名前="☜(ﾟヮﾟ☜)"
+      smile "😁"
+      ノード お名前="☜(ﾟヮﾟ☜)"
     """
         .trim());
     var nodes = KdlDocument([
@@ -725,79 +481,25 @@ smile "😁"
 
   test('node_names', () {
     var doc = parser.parse(r"""
-"!@$@$%Q$%~@!40" "1.2.3" "!!!!!"=#true
-foo123~!@$%^&*.:'|?+ "weeee"
-- 1
+      "!@#$@$%Q#$%~@!40" "1.2.3" "!!!!!"=true
+      foo123~!@#$%^&*.:'|?+ "weeee"
     """
         .trim());
     var nodes = KdlDocument([
-      KdlNode(r"!@$@$%Q$%~@!40",
+      KdlNode(r"!@#$@$%Q#$%~@!40",
           arguments: [KdlString("1.2.3")],
           properties: {"!!!!!": KdlBool(true)}),
-      KdlNode(r"foo123~!@$%^&*.:'|?+", arguments: [KdlString("weeee")]),
-      KdlNode('-', arguments: [KdlInt(1)]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('escaping', () {
-    var doc = parser.parse("""
-node1 "\\u{1f600}"
-node2 "\\n\\t\\r\\\\\\"\\f\\b"
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('node1', arguments: [KdlString('😀')]),
-      KdlNode('node2', arguments: [KdlString("\n\t\r\\\"\f\b")]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('node type', () {
-    var doc = parser.parse("(foo)node");
-    var nodes = KdlDocument([
-      KdlNode('node', type: 'foo'),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('value type', () {
-    var doc = parser.parse('node (foo)"bar"');
-    var nodes = KdlDocument([
-      KdlNode('node', arguments: [KdlString("bar").asType("foo")]),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('property type', () {
-    var doc = parser.parse('node baz=(foo)"bar"');
-    var nodes = KdlDocument([
-      KdlNode('node', properties: {'baz': KdlString("bar").asType("foo")}),
-    ]);
-    expect(doc, equals(nodes));
-  });
-
-  test('child type', () {
-    var doc = parser.parse("""
-node {
-  (foo)bar
-}
-    """
-        .trim());
-    var nodes = KdlDocument([
-      KdlNode('node', children: [
-        KdlNode('bar', type: 'foo'),
-      ]),
+      KdlNode(r"foo123~!@#$%^&*.:'|?+", arguments: [KdlString("weeee")])
     ]);
     expect(doc, equals(nodes));
   });
 
   test('version directive', () {
-    var doc = parser.parse('/- kdl-version 2\nnode foo');
+    var doc = parser.parse('/- kdl-version 1\nnode "foo"');
     expect(doc, isNotNull);
 
     expect(() {
-      parser.parse('/- kdl-version 1\nnode "foo"');
+      parser.parse('/- kdl-version 2\nnode foo');
     }, throwsA(isA<KdlVersionMismatchException>()));
   });
 }
