@@ -1,13 +1,16 @@
 import "../document.dart";
 import "./currency/iso4217_currencies.dart";
 
+/// ISO 4217 currency code.
 class KdlCurrency extends KdlValue<Currency> {
-  KdlCurrency(Currency value, [String? type]) : super(value, type);
+  /// Construct a new `KdlCurrency`
+  KdlCurrency(super.value, [super.type]);
 
-  static call(KdlValue value, [String type = 'currency']) {
-    if (!(value is KdlString)) return null;
-    
-    var currency = Currency.CURRENCIES[value.value];
+  /// Convert a `KdlString` into a `KdlCurrency`
+  static KdlCurrency? convert(KdlValue value, [String type = 'currency']) {
+    if (value is! KdlString) return null;
+
+    var currency = Currency.currencies[value.value];
     if (currency == null) throw "invalid currency: ${value.value}";
 
     return KdlCurrency(currency, type);
