@@ -23,122 +23,134 @@
 
 import 'package:string_scanner/string_scanner.dart';
 
-enum DurationParsingMode {
-  Start,
-  Sign,
-  Date,
-  Time,
+enum _DurationParsingMode {
+  start,
+  sign,
+  date,
+  time,
 }
 
-// Parses a string formatted according to ISO 8601 Duration into the hash.
-//
-// See {ISO 8601}[https://en.wikipedia.org/wiki/ISO_8601#Durations] for more information.
-//
-// This parser allows negative parts to be present in pattern.
-class ISO8601DurationParser {
-  static final PERIOD_OR_COMMA = RegExp('\.|,/');
-  static const PERIOD = '.';
-  static const COMMA = ',';
+/// Parses a string formatted according to ISO 8601 Duration into the hash.
+///
+/// See {ISO 8601}[https://en.wikipedia.org/wiki/ISO_8601#Durations] for more information.
+///
+/// This parser allows negative parts to be present in pattern.
+class Iso8601DurationParser {
+  static final _periodOrComma = RegExp('\\.|,');
+  static const _period = '.';
+  static const _comma = ',';
 
-  static final SIGN_MARKER = RegExp('\\A-|\\+|');
-  static final DATE_MARKER = RegExp('P');
-  static final TIME_MARKER = RegExp('T');
-  static final DATE_COMPONENT = RegExp('(-?\\d+(?:[.,]\\d+)?)(Y|M|D|W)');
-  static final TIME_COMPONENT = RegExp('(-?\\d+(?:[.,]\\d+)?)(H|M|S)');
+  static final _signMarker = RegExp('^-|\\+|');
+  static final _dateMarker = RegExp('P');
+  static final _timeMarker = RegExp('T');
+  static final _dateComponent = RegExp('(-?\\d+(?:[.,]\\d+)?)(Y|M|D|W)');
+  static final _timeComponent = RegExp('(-?\\d+(?:[.,]\\d+)?)(H|M|S)');
 
-  static const DATE_TO_PART = { 'Y': 'years', 'M': 'months', 'W': 'weeks', 'D': 'days' };
-  static const TIME_TO_PART = { 'H': 'hours', 'M': 'minutes', 'S': 'seconds' };
+  static const _dateToPart = {
+    'Y': 'years',
+    'M': 'months',
+    'W': 'weeks',
+    'D': 'days'
+  };
+  static const _timeToPart = {'H': 'hours', 'M': 'minutes', 'S': 'seconds'};
 
-  static const DATE_COMPONENTS = ['years', 'months', 'days'];
-  static const TIME_COMPONENTS = ['hours', 'minutes', 'seconds'];
+  static const _dateComponents = ['years', 'months', 'days'];
+  static const _timeComponents = ['hours', 'minutes', 'seconds'];
 
-  Map<String, num> parts;
-  StringScanner scanner;
-  DurationParsingMode mode;
-  int sign;
+  final Map<String, num> _parts;
+  final StringScanner _scanner;
+  _DurationParsingMode _mode;
+  int _sign;
 
-  ISO8601DurationParser(String string) :
-    scanner = StringScanner(string),
-    parts = {},
-    mode = DurationParsingMode.Start,
-    sign = 1;
+  /// Construct a new parser to parse the given string
+  Iso8601DurationParser(String string)
+      : _scanner = StringScanner(string),
+        _parts = {},
+        _mode = _DurationParsingMode.start,
+        _sign = 1;
 
-  parse() {
+  /// Parse the string into a map of duration parts
+  Map<String, num> parse() {
     while (!_isFinished()) {
-      switch(mode) {
-      case DurationParsingMode.Start:
-        if (_scan(SIGN_MARKER)) {
-          sign = scanner.lastMatch![0] == '-' ? -1 : 1;
-          mode = DurationParsingMode.Sign;
-        } else {
-          _raiseParsingError();
-        }
-        break;
-      case DurationParsingMode.Sign:
-        if (_scan(DATE_MARKER)) {
-          mode = DurationParsingMode.Date;
-        } else {
-          _raiseParsingError();
-        }
-        break;
-      case DurationParsingMode.Date:
-        if (_scan(TIME_MARKER)) {
-          mode = DurationParsingMode.Time;
-        } else if (_scan(DATE_COMPONENT)) {
-          parts[DATE_TO_PART[scanner.lastMatch![2]!]!] = _number() * sign;
-        } else {
-          _raiseParsingError();
-        }
-        break;
-      case DurationParsingMode.Time:
-        if (_scan(TIME_COMPONENT)) {
-          parts[TIME_TO_PART[scanner.lastMatch![2]!]!] = _number() * sign;
-        } else {
-          _raiseParsingError();
-        }
-        break;
+      switch (_mode) {
+        case _DurationParsingMode.start:
+          if (_scan(_signMarker)) {
+            _sign = _scanner.lastMatch![0] == '-' ? -1 : 1;
+            _mode = _DurationParsingMode.sign;
+          } else {
+            _raiseParsingError();
+          }
+          break;
+        case _DurationParsingMode.sign:
+          if (_scan(_dateMarker)) {
+            _mode = _DurationParsingMode.date;
+          } else {
+            _raiseParsingError();
+          }
+          break;
+        case _DurationParsingMode.date:
+          if (_scan(_timeMarker)) {
+            _mode = _DurationParsingMode.time;
+          } else if (_scan(_dateComponent)) {
+            _parts[_dateToPart[_scanner.lastMatch![2]!]!] = _number() * _sign;
+          } else {
+            _raiseParsingError();
+          }
+          break;
+        case _DurationParsingMode.time:
+          if (_scan(_timeComponent)) {
+            _parts[_timeToPart[_scanner.lastMatch![2]!]!] = _number() * _sign;
+          } else {
+            _raiseParsingError();
+          }
+          break;
       }
     }
 
     _validate();
-    return parts;
+    return _parts;
   }
 
   bool _isFinished() {
-    return scanner.rest.length == 0;
+    return _scanner.rest.isEmpty;
   }
 
   // Parses number which can be a float with either comma or period.
   num _number() {
-    return (PERIOD_OR_COMMA.hasMatch(scanner.lastMatch![1]!)) ?
-      double.parse(scanner.lastMatch![1]!.replaceAll(COMMA, PERIOD)) :
-      int.parse(scanner.lastMatch![1]!);
+    return (_periodOrComma.hasMatch(_scanner.lastMatch![1]!))
+        ? double.parse(_scanner.lastMatch![1]!.replaceAll(_comma, _period))
+        : int.parse(_scanner.lastMatch![1]!);
   }
 
   bool _scan(pattern) {
-    return scanner.scan(pattern);
+    return _scanner.scan(pattern);
   }
 
   void _raiseParsingError([String? reason]) {
-    throw "Invalid ISO 8601 duration: ${scanner.string} $reason".trim();
+    throw "Invalid ISO 8601 duration: ${_scanner.string} $reason".trim();
   }
 
   // Checks for various semantic errors as stated in ISO 8601 standard.
   bool _validate() {
-    if (parts.isEmpty) _raiseParsingError('is empty duration');
+    if (_parts.isEmpty) _raiseParsingError('is empty duration');
 
     // Mixing any of Y, M, D with W is invalid.
-    if (parts.containsKey('weeks') && DATE_COMPONENTS.any((e) => parts.containsKey(e))) {
+    if (_parts.containsKey('weeks') &&
+        _dateComponents.any((e) => _parts.containsKey(e))) {
       _raiseParsingError('mixing weeks with other date parts not allowed');
     }
 
     // Specifying an empty T part is invalid.
-    if (mode == DurationParsingMode.Time && !TIME_COMPONENTS.any((e) => parts.containsKey(e))) {
+    if (_mode == _DurationParsingMode.time &&
+        !_timeComponents.any((e) => _parts.containsKey(e))) {
       _raiseParsingError('time part marker is present but time part is empty');
     }
 
-    var fractions = parts.values.where((a) => a != 0).where((a) => (a % 1) != 0);
-    if (!fractions.isEmpty && !(fractions.length == 1 && fractions.last == parts.values.where((a) => a != 0).last)) {
+    var fractions =
+        _parts.values.where((a) => a != 0).where((a) => (a % 1) != 0);
+    if (fractions.isNotEmpty &&
+        !(fractions.length == 1 &&
+            fractions.last == _parts.values.where((a) => a != 0).last)) {
       _raiseParsingError('(only last part can be fractional)');
     }
 
