@@ -1,10 +1,13 @@
 import "../document.dart";
 
+/// Regular expression.
 class KdlRegex extends KdlValue<RegExp> {
-  KdlRegex(RegExp value, [String? type]) : super(value, type);
+  /// Construct a new `KdlRegex`
+  KdlRegex(super.value, [super.type]);
 
-  static call(KdlValue value, [String type = 'regex']) {
-    if (!(value is KdlString)) return null;
+  /// Convert a `KdlString` into a `KdlRegex`
+  static KdlRegex? convert(KdlValue value, [String type = 'regex']) {
+    if (value is! KdlString) return null;
 
     return KdlRegex(RegExp(value.value), type);
   }
