@@ -6,9 +6,9 @@ import 'package:kdl/src/types/decimal.dart';
 
 void main() {
   test('decimal', () {
-    expect(KdlDecimal.call(KdlString('10000000000000')).value,
+    expect(KdlDecimal.convert(KdlString('10000000000000'))!.value,
       equals(BigDecimal.parse('10000000000000')));
 
-    expect(() => KdlDecimal.call(KdlString('not a decimal')), throwsA(anything));
+    expect(() => KdlDecimal.convert(KdlString('not a decimal')), throwsA(anything));
   });
 }
