@@ -5,12 +5,12 @@ import 'package:kdl/src/types/email.dart';
 
 void main() {
   test('email', () {
-    var value = KdlEmail.call(KdlString('danielle@example.com'));
+    var value = KdlEmail.convert(KdlString('danielle@example.com'))!;
     expect(value.value, equals('danielle@example.com'));
     expect(value.local, equals('danielle'));
     expect(value.domain, equals('example.com'));
 
-    expect(() => KdlEmail.call(KdlString('not an email')), throwsA(anything));
+    expect(() => KdlEmail.convert(KdlString('not an email')), throwsA(anything));
   });
 
   var validEmails = [
@@ -34,7 +34,7 @@ void main() {
 
   test('valid emails', () {
     for (var testCase in validEmails) {
-      var value = KdlEmail.call(KdlString(testCase[0]));
+      var value = KdlEmail.convert(KdlString(testCase[0]))!;
       expect(value.value, equals(testCase[0]));
       expect(value.local, equals(testCase[1]));
       expect(value.domain, equals(testCase[2]));
@@ -44,10 +44,10 @@ void main() {
   var invalidEmails = [
     'Abc.example.com',
     'A@b@c@example.com',
-    'a"b(c)d,e:f;g<h>i[j\k]l@example.com',
+    'a"b(c)d,e:f;g<h>i[j\\k]l@example.com',
     'just"not"right@example.com',
-    'this is"not\allowed@example.com',
-    'this\ still\"not\\allowed@example.com',
+    'this is"not\\allowed@example.com',
+    'this\\ still\\"not\\allowed@example.com',
     '1234567890123456789012345678901234567890123456789012345678901234+x@example.com',
     '-some-user-@-example-.com',
     'QA🦄CHOCOLATE🌈@test.com',
@@ -55,24 +55,24 @@ void main() {
 
   test('invalid emails', () {
     for (var email in invalidEmails) {
-      expect(() => KdlEmail.call(KdlString(email)), throwsA(anything));
+      expect(() => KdlEmail.convert(KdlString(email)), throwsA(anything));
     }
   });
 
   test('idn email', () {
-    var value = KdlIDNEmail.call(KdlString('🌈@xn--9ckb.com'));
+    var value = KdlIdnEmail.convert(KdlString('🌈@xn--9ckb.com'))!;
     expect(value.value, equals('🌈@xn--9ckb.com'));
     expect(value.unicodeValue, equals('🌈@ツッ.com'));
     expect(value.local, equals('🌈'));
     expect(value.unicodeDomain, equals('ツッ.com'));
     expect(value.domain, equals('xn--9ckb.com'));
-    value = KdlIDNEmail.call(KdlString('🌈@ツッ.com'));
+    value = KdlIdnEmail.convert(KdlString('🌈@ツッ.com'))!;
     expect(value.value, equals('🌈@xn--9ckb.com'));
     expect(value.unicodeValue, equals('🌈@ツッ.com'));
     expect(value.local, equals('🌈'));
     expect(value.unicodeDomain, equals('ツッ.com'));
     expect(value.domain, equals('xn--9ckb.com'));
 
-    expect(() => KdlIDNEmail.call(KdlString('not an email')), throwsA(anything));
+    expect(() => KdlIdnEmail.convert(KdlString('not an email')), throwsA(anything));
   });
 }
