@@ -5,34 +5,42 @@ import 'package:kdl/src/types/hostname.dart';
 
 void main() {
   test('hostname', () {
-    expect(KdlHostname.call(KdlString('www.example.com')).value,
-      equals('www.example.com'));
+    expect(KdlHostname.convert(KdlString('www.example.com'))!.value,
+        equals('www.example.com'));
 
     // 63 a's
-    var maxPartLength = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.com';
-    expect(KdlHostname.call(KdlString(maxPartLength)).value, equals(maxPartLength));
+    var maxPartLength =
+        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.com';
+    expect(KdlHostname.convert(KdlString(maxPartLength))!.value,
+        equals(maxPartLength));
 
-    var maxLength = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.' +
-      'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.' +
-      'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.' +
-      'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
-    expect(KdlHostname.call(KdlString(maxLength)).value, equals(maxLength));
+    var maxLength =
+        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.'
+        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.'
+        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa.'
+        'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa';
+    expect(KdlHostname.convert(KdlString(maxLength))?.value, equals(maxLength));
 
-    expect(() => KdlHostname.call(KdlString('not a hostname')), throwsA(anything));
-    expect(() => KdlHostname.call(KdlString('-start-with-a-dash.com')), throwsA(anything));
-    expect(() => KdlHostname.call(KdlString('a' + maxPartLength)), throwsA(anything));
-    expect(() => KdlHostname.call(KdlString(maxLength + 'a')), throwsA(anything));
+    expect(
+        () => KdlHostname.convert(KdlString('not a hostname')), throwsA(anything));
+    expect(() => KdlHostname.convert(KdlString('-start-with-a-dash.com')),
+        throwsA(anything));
+    expect(() => KdlHostname.convert(KdlString('a$maxPartLength')),
+        throwsA(anything));
+    expect(
+        () => KdlHostname.convert(KdlString('${maxLength}a')), throwsA(anything));
   });
 
   test('idn hostname', () {
-    var value = KdlIDNHostname.call(KdlString('xn--bcher-kva.example'));
+    var value = KdlIdnHostname.convert(KdlString('xn--bcher-kva.example'))!;
     expect(value.value, equals('xn--bcher-kva.example'));
     expect(value.unicodeValue, equals('bücher.example'));
 
-    value = KdlIDNHostname.call(KdlString('bücher.example'));
+    value = KdlIdnHostname.convert(KdlString('bücher.example'))!;
     expect(value.value, equals('xn--bcher-kva.example'));
     expect(value.unicodeValue, equals('bücher.example'));
 
-    expect(() => KdlIDNHostname.call(KdlString('not a hostname')), throwsA(anything));
+    expect(() => KdlIdnHostname.convert(KdlString('not a hostname')),
+        throwsA(anything));
   });
 }
