@@ -5,16 +5,16 @@ import 'package:kdl/src/types/url.dart';
 
 void main() {
   test('url', () {
-    expect(KdlURL.call(KdlString('https://www.example.com/foo/bar')).value,
+    expect(KdlUrl.convert(KdlString('https://www.example.com/foo/bar'))!.value,
       equals(Uri.parse('https://www.example.com/foo/bar')));
 
-    expect(() => KdlURL.call(KdlString('/reference/to/something')), throwsA(anything));
+    expect(() => KdlUrl.convert(KdlString('/reference/to/something')), throwsA(anything));
   });
 
   test('url reference', () {
-    expect(KdlURLReference.call(KdlString('https://www.example.com/foo/bar')).value,
+    expect(KdlUrlReference.convert(KdlString('https://www.example.com/foo/bar'))!.value,
       equals(Uri.parse('https://www.example.com/foo/bar')));
-    expect(KdlURLReference.call(KdlString('/foo/bar')).value,
+    expect(KdlUrlReference.convert(KdlString('/foo/bar'))!.value,
       equals(Uri.parse('/foo/bar')));
   });
 }
