@@ -1,36 +1,21 @@
-
 import 'package:test/test.dart';
 
 import 'package:kdl/kdl.dart';
-import 'package:kdl/src/document.dart';
 import 'package:kdl/src/parser.dart';
-import 'package:kdl/src/types/date_time.dart';
-import 'package:kdl/src/types/duration.dart';
-import 'package:kdl/src/types/decimal.dart';
-import 'package:kdl/src/types/currency.dart';
-import 'package:kdl/src/types/country.dart';
-import 'package:kdl/src/types/email.dart';
-import 'package:kdl/src/types/hostname.dart';
-import 'package:kdl/src/types/ip.dart';
-import 'package:kdl/src/types/url.dart';
-import 'package:kdl/src/types/irl.dart';
-import 'package:kdl/src/types/url_template.dart';
-import 'package:kdl/src/types/uuid.dart';
-import 'package:kdl/src/types/regex.dart';
-import 'package:kdl/src/types/base64.dart';
 
 class Foo extends KdlValue<String> {
-  Foo(String value, [String? type]) : super(value, type);
+  Foo(super.value, [super.type]);
 }
 
 class Bar extends KdlNode {
-  Bar(KdlNode node, [String? type]) : super(
-    node.name,
-    arguments: node.arguments,
-    properties: node.properties,
-    children: node.children,
-    type: type,
-  );
+  Bar(KdlNode node, [String? type])
+      : super(
+          node.name,
+          arguments: node.arguments,
+          properties: node.properties,
+          children: node.children,
+          type: type,
+        );
 }
 
 void main() {
@@ -66,7 +51,8 @@ node (date-time)"2021-01-01T12:12:12" \\
      (uuid)"f81d4fae-7dec-11d0-a765-00a0c91e6bf6" \\
      (regex)"asdf" \\
      (base64)"U2VuZCByZWluZm9yY2VtZW50cw=="
-""".trim());
+"""
+        .trim());
 
     var i = 0;
     expect(doc.nodes[0].arguments[i++], isA<KdlDateTime>());
@@ -79,36 +65,40 @@ node (date-time)"2021-01-01T12:12:12" \\
     expect(doc.nodes[0].arguments[i++], isA<KdlCountry3>());
     expect(doc.nodes[0].arguments[i++], isA<KdlCountrySubdivision>());
     expect(doc.nodes[0].arguments[i++], isA<KdlEmail>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlIDNEmail>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlIdnEmail>());
     expect(doc.nodes[0].arguments[i++], isA<KdlHostname>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlIDNHostname>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlIdnHostname>());
     expect(doc.nodes[0].arguments[i++], isA<KdlIPV4>());
     expect(doc.nodes[0].arguments[i++], isA<KdlIPV6>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlURL>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlURLReference>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlUrl>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlUrlReference>());
     expect(doc.nodes[0].arguments[i++], isA<KdlIRL>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlIRLReference>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlURLTemplate>());
-    expect(doc.nodes[0].arguments[i++], isA<KdlUUID>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlIrlReference>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlUrlTemplate>());
+    expect(doc.nodes[0].arguments[i++], isA<KdlUuid>());
     expect(doc.nodes[0].arguments[i++], isA<KdlRegex>());
     expect(doc.nodes[0].arguments[i++], isA<KdlBase64>());
   });
 
   test('custom types', () {
-    var parsers = {
-      'foo': (value, type) {
-        if (!(value is KdlValue)) return null;
+    var customValues = {
+      'foo': (KdlValue value, String type) {
         return Foo(value.value, type);
-      },
-      'bar': (node, type) {
-        if (!(node is KdlNode)) return null;
+      }
+    };
+    var customNodes = {
+      'bar': (KdlNode node, type) {
         return Bar(node, type);
       },
     };
-    var doc = Kdl.parseDocument("""
+    var doc = KdlDocument.parse(
+        """
 (bar)barnode (foo)"foovalue"
 (foo)foonode (bar)"barvalue"
-""".trim(), typeParsers: parsers);
+"""
+            .trim(),
+        valueTypes: customValues,
+        nodeTypes: customNodes);
 
     expect(doc, isNotNull);
     expect(doc.nodes[0], isA<Bar>());
@@ -118,9 +108,12 @@ node (date-time)"2021-01-01T12:12:12" \\
   });
 
   test('parse false', () {
-    var doc = Kdl.parseDocument("""
+    var doc = KdlDocument.parse(
+        """
 node (date-time)"2021-01-01T12:12:12"
-    """.trim(), parseTypes: false);
+    """
+            .trim(),
+        parseTypes: false);
 
     expect(doc, isNotNull);
     expect(doc.nodes[0].arguments[0], isA<KdlString>());
